+++ src/com/intellij/idea/plugin/hybris/settings/HybrisProjectSettings.kt
@@ -1,5 +1,5 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellire IDEA.
  * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify